refactor(rollup-plugin-polyfills-loader): migrate tests to node:test
Replace mocha globals and chai assertions with node:test and
node:assert/strict across 3 test files.

- `import type { OutputChunk, OutputAsset, RollupOptions, OutputOptions }` (type-only)
- `import { polyfillsLoader } from '../../dist/index.js'` (named, not default)
- `../../src/*.js` -> `../../dist/*.js`
- `expect(x).to.eql(y)` -> `assert.deepEqual(x, y)`
- `expect(x).to.equal(y)` -> `assert.equal(x, y)`
- `expect(action).to.throw()` -> `assert.throws(action)`
- `expect(x).to.exist` -> `assert.ok(x)`
- `this.timeout(5000)` -> `{ timeout: 5000 }` on describe
- `__dirname` -> `import.meta.dirname`
- Removed `test:update-snapshots` script
- Add `--experimental-strip-types` for CI compat
