+++ packages/rollup-plugin-polyfills-loader/package.json
@@ -25,9 +25,8 @@
     "node": ">=22.0.0"
   },
   "scripts": {
-    "test:node": "mocha test/**/*.test.ts --require ts-node/register --reporter dot",
-    "test:update-snapshots": "mocha test/**/*.test.ts --require ts-node/register --update-snapshots",
-    "test:watch": "mocha test/**/*.test.ts --require ts-node/register --watch --watch-files src,test"
+    "test:node": "node --experimental-strip-types --test --test-force-exit test/**/*.test.ts",
+    "test:watch": "node --experimental-strip-types --test --test-force-exit --watch test/**/*.test.ts"
   },
   "files": [
     "*.d.ts",
