+++ packages/rollup-plugin-polyfills-loader/test/src/rollupPluginPolyfillsLoader.test.ts
@@ -1,14 +1,16 @@
 /* eslint-disable no-await-in-loop */
-import { OutputChunk, rollup, OutputAsset, RollupOptions, OutputOptions } from 'rollup';
-import { expect } from 'chai';
+import { describe, it } from 'node:test';
+import assert from 'node:assert/strict';
+import { rollup } from 'rollup';
+import type { OutputChunk, OutputAsset, RollupOptions, OutputOptions } from 'rollup';
 import fs from 'fs';
 import path from 'path';
 import { rollupPluginHTML as html } from '@web/rollup-plugin-html';
-import polyfillsLoader from '../../src/index.js';
+import { polyfillsLoader } from '../../dist/index.js';
 
 type Output = (OutputChunk | OutputAsset)[];
 
-const relativeUrl = `./${path.relative(process.cwd(), path.join(__dirname, '..'))}`;
+const relativeUrl = `./${path.relative(process.cwd(), path.join(import.meta.dirname, '..'))}`;
 
 const updateSnapshots = process.argv.includes('--update-snapshots');
 
@@ -26,7 +28,7 @@ interface SnapshotArgs {
 }
 
 async function testSnapshot({ name, fileName, inputOptions, outputOptions }: SnapshotArgs) {
-  const snapshotPath = path.join(__dirname, '..', 'snapshots', `${name}.html`);
+  const snapshotPath = path.join(import.meta.dirname, '..', 'snapshots', `${name}.html`);
   const bundle = await rollup(inputOptions);
   let output;
   for (const outputConfig of outputOptions) {
@@ -42,8 +44,7 @@ async function testSnapshot({ name, fileName, inputOptions, outputOptions }: Sna
     fs.writeFileSync(snapshotPath, file.source, 'utf-8');
   } else {
     const snapshot = fs.readFileSync(snapshotPath, 'utf-8');
-    expect(file.source.trim()).to.equal(snapshot.trim());
-    // expect(file.source.replace(/\s/g, '')).to.equal(snapshot.replace(/\s/g, ''));
+    assert.equal(file.source.trim(), snapshot.trim());
   }
   return output;
 }
@@ -55,10 +56,7 @@ const defaultOutputOptions: OutputOptions[] = [
   },
 ];
 
-describe('rollup-plugin-polyfills-loader', function describe() {
-  // bootup of the first test can take a long time in CI to load all the polyfills
-  this.timeout(5000);
-
+describe('rollup-plugin-polyfills-loader', { timeout: 5000 }, () => {
   it('can inject a polyfills loader with a single output', async () => {
     const inputOptions: RollupOptions = {
       plugins: [
@@ -225,8 +223,8 @@ describe('rollup-plugin-polyfills-loader', function describe() {
       outputOptions: defaultOutputOptions,
     });
 
-    expect(output.find(o => o.fileName.startsWith('polyfills/webcomponents'))).to.exist;
-    expect(output.find(o => o.fileName.startsWith('polyfills/fetch'))).to.exist;
+    assert.ok(output.find(o => o.fileName.startsWith('polyfills/webcomponents')));
+    assert.ok(output.find(o => o.fileName.startsWith('polyfills/fetch')));
   });
 
   it('can inject with multiple build outputs', async () => {
