+++ packages/rollup-plugin-polyfills-loader/test/src/utils.test.ts
@@ -1,53 +1,59 @@
-import { expect } from 'chai';
+import { describe, it } from 'node:test';
+import assert from 'node:assert/strict';
 import { fileTypes } from '@web/polyfills-loader';
-import { shouldInjectLoader } from '../../src/utils.js';
+import { shouldInjectLoader } from '../../dist/utils.js';
 
 describe('shouldInjectLoader', () => {
   it('returns true when modern contains non-module or script', () => {
-    expect(
+    assert.equal(
       shouldInjectLoader({
         modern: { files: [{ type: fileTypes.SYSTEMJS, path: '' }] },
       }),
-    ).to.equal(true);
+      true,
+    );
   });
 
   it('returns true when there are legacy files', () => {
-    expect(
+    assert.equal(
       shouldInjectLoader({
         modern: { files: [{ type: fileTypes.MODULE, path: '' }] },
         legacy: [{ test: '', files: [{ type: fileTypes.SYSTEMJS, path: '' }] }],
       }),
-    ).to.equal(true);
+      true,
+    );
   });
 
   it('returns true when there are polyfills', () => {
-    expect(
+    assert.equal(
       shouldInjectLoader({
         modern: { files: [{ type: fileTypes.MODULE, path: '' }] },
         polyfills: {
           fetch: true,
         },
       }),
-    ).to.equal(true);
+      true,
+    );
 
-    expect(
+    assert.equal(
       shouldInjectLoader({
         modern: { files: [{ type: fileTypes.MODULE, path: '' }] },
         polyfills: {
           regeneratorRuntime: 'always',
         },
       }),
-    ).to.equal(true);
+      true,
+    );
   });
 
   it('returns true when there are custom polyfills', () => {
-    expect(
+    assert.equal(
       shouldInjectLoader({
         modern: { files: [{ type: fileTypes.MODULE, path: '' }] },
         polyfills: {
           custom: [{ test: '', path: '', name: '' }],
         },
       }),
-    ).to.equal(true);
+      true,
+    );
   });
 });
